backend/requirements.txt
# AI-Powered News CMS - Backend Dependencies
# Python 3.11+

# FastAPI Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Pydantic v2 for data validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# MongoDB Driver (Motor - Async)
motor>=3.6.0
pymongo>=4.10.0

# AWS SDK for Secrets Manager
boto3>=1.35.0
botocore>=1.35.0

# LLM Provider SDKs
openai>=1.57.0
anthropic>=0.42.0
google-generativeai>=0.8.0

# HTTP Client
httpx>=0.28.0
requests>=2.32.0

# Environment Variables
python-dotenv==1.0.1

# CORS Support
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
pytest-httpx>=0.35.0

# Linting and Formatting
flake8>=7.1.0
black>=24.10.0
isort>=5.13.0
mypy>=1.13.0

# Utilities
python-dateutil>=2.9.0
pytz>=2024.1
